Desktop/bachat-ai-deployment-project/bachat-ai/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter>=3.5.0
python-dotenv==1.0.0
supabase>=2.4.0
pdfplumber==0.10.3
PyMuPDF==1.23.8
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
openpyxl==3.1.2
reportlab==4.0.8
google-generativeai>=0.8.0
gunicorn==21.2.0
waitress>=2.1.0
requests>=2.31.0
websockets>=14.0
psycopg2-binary>=2.9.9
pg8000>=1.30.3
